requirements.txt
requests
openai
python-dotenv
beautifulsoup4
arxiv
PyPDF2
